ABI-244 - add encode / decode User ID function in injection object
